5052 - заменила id школы на Long / добавила метод edit

+++ src/main/java/com/kataacademy/schoolportal/common/repository/schooattribute/SchoolRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface SchoolRepository extends JpaRepository<School, Integer> {
+public interface SchoolRepository extends JpaRepository<School, Long> {
 
 }
